Switch backspace & enter, level3 on windows key.
Enter on the right and backspace on the left is more intuitive. This
forces the level3 switch on the left hand side to the windows key.

Also, parentheses, brackets and braces are now on fg, rt and vb
respectively. This makes for another intuitive cluster.

PageUp & PageDown are on xc now.

Time will tell if this is a sensible configuration. I wanted to try this
asap so that my muscle memory doesn't get too used to backspace on the
right.